+++ example/materialBall.js
@@ -12,6 +12,7 @@ let renderer, controls, sceneInfo, ptRenderer, activeCamera, blitQuad, denoiseQu
 let perspectiveCamera, orthoCamera, equirectCamera;
 let envMap, envMapGenerator, scene;
 let samplesEl;
+let PT_PROGRAM_ID;
 
 const orthoWidth = 5;
 
@@ -39,6 +40,7 @@ const params = {
 		specularColor: '#ffffff',
 		specularIntensity: 1.0,
 		matte: false,
+		flatShading: false,
 		castShadow: true,
 	},
 	material2: {
@@ -63,6 +65,7 @@ const params = {
 		specularColor: '#ffffff',
 		specularIntensity: 1.0,
 		matte: false,
+		flatShading: false,
 		castShadow: true,
 	},
 	material3: {
@@ -449,6 +452,7 @@ async function init() {
 	matFolder1.addColor( params.material1, 'specularColor' ).onChange( reset );
 	matFolder1.add( params.material1, 'specularIntensity', 0.0, 1.0 ).onChange( reset );
 	matFolder1.add( params.material1, 'matte' ).onChange( reset );
+	matFolder1.add( params.material1, 'flatShading' ).onChange( reset );
 	matFolder1.add( params.material1, 'castShadow' ).onChange( reset );
 	matFolder1.close();
 
@@ -474,6 +478,7 @@ async function init() {
 	matFolder2.addColor( params.material2, 'specularColor' ).onChange( reset );
 	matFolder2.add( params.material2, 'specularIntensity', 0.0, 1.0 ).onChange( reset );
 	matFolder2.add( params.material2, 'matte' ).onChange( reset );
+	matFolder2.add( params.material2, 'flatShading' ).onChange( reset );
 	matFolder2.add( params.material2, 'castShadow' ).onChange( reset );
 	matFolder2.close();
 
@@ -596,6 +601,7 @@ function animate() {
 	m1.specularColor.set( params.material1.specularColor ).convertSRGBToLinear();
 	m1.specularIntensity = params.material1.specularIntensity;
 	m1.transparent = m1.opacity < 1;
+	m1.flatShading = params.material1.flatShading;
 
 	const m2 = materials[ 1 ];
 	m2.color.set( params.material2.color ).convertSRGBToLinear();
@@ -618,6 +624,7 @@ function animate() {
 	m2.specularColor.set( params.material2.specularColor ).convertSRGBToLinear();
 	m2.specularIntensity = params.material2.specularIntensity;
 	m2.transparent = m2.opacity < 1;
+	m2.flatShading = params.material2.flatShading;
 
 	const m3 = materials[ 2 ];
 	m3.color.set( params.material3.color ).convertSRGBToLinear();
@@ -651,9 +658,40 @@ function animate() {
 
 	}
 
+	// Get the path tracing shader id. It will be the next program compiled and used here.
+	if ( PT_PROGRAM_ID === undefined ) {
+
+		PT_PROGRAM_ID = renderer.info.programs.length;
+
+	}
+
 	for ( let i = 0, l = params.samplesPerFrame; i < l; i ++ ) {
 
 		ptRenderer.update();
+
+	}
+
+	if ( ! window.DEBUG_SHADERS ) {
+
+		const gl = renderer.getContext();
+		const DEBUG_SHADERS = {};
+
+		const programs = renderer.info.programs;
+		const ptProgram = programs[ PT_PROGRAM_ID ];
+		DEBUG_SHADERS.context = gl;
+
+		DEBUG_SHADERS.vertexShader = ptProgram.vertexShader;
+		DEBUG_SHADERS.fragmentShader = ptProgram.fragmentShader;
+
+		DEBUG_SHADERS.vertexShaderSource = gl.getShaderSource( ptProgram.vertexShader );
+		DEBUG_SHADERS.fragmentShaderSource = gl.getShaderSource( ptProgram.fragmentShader );
+
+		const shaderDebugExt = gl.getExtension( 'WEBGL_debug_shaders' );
+		DEBUG_SHADERS.debugVertexShaderSource = shaderDebugExt.getTranslatedShaderSource( ptProgram.vertexShader );
+		DEBUG_SHADERS.debugFragmentShaderSource = shaderDebugExt.getTranslatedShaderSource( ptProgram.fragmentShader );
+
+		window.DEBUG_SHADERS = DEBUG_SHADERS;
+		console.log( 'Original and translated debug shaders added to window.DEBUG_SHADERS object.' );
 
 	}