+++ src/shader/shaderRandFunctions.js
@@ -1,0 +1,57 @@
+export const shaderRandFunctions = /* glsl */`
+
+	// https://www.shadertoy.com/view/wltcRS
+	uvec4 WHITE_NOISE_SEED;
+
+	void rng_initialize( vec2 p, int frame ) {
+
+		// white noise seed
+		WHITE_NOISE_SEED = uvec4( p, uint( frame ), uint( p.x ) + uint( p.y ) );
+
+	}
+
+	// https://www.pcg-random.org/
+	void pcg4d( inout uvec4 v ) {
+
+		v = v * 1664525u + 1013904223u;
+		v.x += v.y * v.w;
+		v.y += v.z * v.x;
+		v.z += v.x * v.y;
+		v.w += v.y * v.z;
+		v = v ^ ( v >> 16u );
+		v.x += v.y*v.w;
+		v.y += v.z*v.x;
+		v.z += v.x*v.y;
+		v.w += v.y*v.z;
+
+	}
+
+	// returns [ 0, 1 ]
+	float rand() {
+
+		pcg4d( WHITE_NOISE_SEED );
+		return float( WHITE_NOISE_SEED.x ) / float( 0xffffffffu );
+
+	}
+
+	vec2 rand2() {
+
+		pcg4d( WHITE_NOISE_SEED );
+		return vec2( WHITE_NOISE_SEED.xy ) / float(0xffffffffu);
+
+	}
+
+	vec3 rand3() {
+
+		pcg4d( WHITE_NOISE_SEED );
+		return vec3( WHITE_NOISE_SEED.xyz ) / float( 0xffffffffu );
+
+	}
+
+	vec4 rand4() {
+
+		pcg4d( WHITE_NOISE_SEED );
+		return vec4( WHITE_NOISE_SEED ) / float( 0xffffffffu );
+
+	}
+`;