+++ src/uniforms/utils.js
@@ -1,3 +1,11 @@
+// we must hash the texture to determine uniqueness using the encoding, as well, because the
+// when rendering each texture to the texture array they must have a consistent color space.
+export function getTextureHash( t ) {
+
+	return `${ t.source.uuid }:${ t.encoding }`;
+
+}
+
 // reduce the set of textures to just those with a unique source while retaining
 // the order of the textures.
 export function reduceTexturesToUniqueSources( textures ) {
@@ -7,9 +15,10 @@ export function reduceTexturesToUniqueSources( textures ) {
 	for ( let i = 0, l = textures.length; i < l; i ++ ) {
 
 		const tex = textures[ i ];
-		if ( ! sourceSet.has( tex.source ) ) {
+		const hash = getTextureHash( tex );
+		if ( ! sourceSet.has( hash ) ) {
 
-			sourceSet.add( tex.source );
+			sourceSet.add( hash );
 			result.push( tex );
 
 		}